pkg/storage: Update the last test to work without profile trees
