+++ pkg/storage/series_iterator_range_test.go
@@ -19,6 +19,7 @@ import (
 	"time"
 
 	"github.com/google/uuid"
+	"github.com/parca-dev/parca/pkg/storage/chunkenc"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/stretchr/testify/require"
 )
@@ -65,4 +66,48 @@ func TestMemRangeSeries_Iterator(t *testing.T) {
 
 	require.NoError(t, it.Err())
 	require.Equal(t, int64(421), seen) // 421 would be seen next but 420 was the last value.
+}
+
+func TestGetIndexRange(t *testing.T) {
+	c := chunkenc.FromValuesDelta(2, 4, 6, 7, 8)
+
+	start, end, err := getIndexRange(c.Iterator(nil), 5, 1, 9)
+	require.NoError(t, err)
+	require.Equal(t, uint64(0), start)
+	require.Equal(t, uint64(5), end)
+
+	start, end, err = getIndexRange(c.Iterator(nil), 5, 2, 9)
+	require.NoError(t, err)
+	require.Equal(t, uint64(0), start)
+	require.Equal(t, uint64(5), end)
+
+	start, end, err = getIndexRange(c.Iterator(nil), 5, 3, 6)
+	require.NoError(t, err)
+	require.Equal(t, uint64(1), start)
+	require.Equal(t, uint64(3), end)
+
+	start, end, err = getIndexRange(c.Iterator(nil), 5, 3, 7)
+	require.NoError(t, err)
+	require.Equal(t, uint64(1), start)
+	require.Equal(t, uint64(4), end)
+
+	start, end, err = getIndexRange(c.Iterator(nil), 5, 3, 8)
+	require.NoError(t, err)
+	require.Equal(t, uint64(1), start)
+	require.Equal(t, uint64(5), end)
+
+	start, end, err = getIndexRange(c.Iterator(nil), 5, 3, 9)
+	require.NoError(t, err)
+	require.Equal(t, uint64(1), start)
+	require.Equal(t, uint64(5), end)
+
+	start, end, err = getIndexRange(c.Iterator(nil), 5, 5, 7)
+	require.NoError(t, err)
+	require.Equal(t, uint64(2), start)
+	require.Equal(t, uint64(4), end)
+
+	start, end, err = getIndexRange(NewMultiChunkIterator([]chunkenc.Chunk{c}), 123, 1, 12)
+	require.NoError(t, err)
+	require.Equal(t, uint64(0), start)
+	require.Equal(t, uint64(5), end)
 }
